+++ frog-utils/src/index.js
@@ -48,6 +48,10 @@ export const ReactiveRichText = isBrowser
   ? require('./ReactiveRichText').default // eslint-disable-line global-require
   : () => <p>Node</p>; // React component to make Flow happy, will never be shown
 
+export const LearningItemBlot = isBrowser
+? require('./LearningItemBlot').default // eslint-disable-line global-require
+: () => <p>Node</p>; // React component to make Flow happy, will never be shown
+
 export {
   hideConditional,
   calculateHides,
